Create separate tabs for station and path inputs
