+++ client/.eslintrc.json
@@ -23,6 +23,7 @@
       }
     ],
     "react/react-in-jsx-scope": "off",
+    "react/jsx-props-no-spreading": "off",
     "no-console": ["warn", { "allow": ["warn", "error"] }]
   }
 }
